test(alignment): separate previous-offset policy coverage
